functions/create_complete_movie/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["sey323 <sey323_nm@yahoo.co.jp>"]
readme = "README.md"
packages = [{include = "create_complete_movie"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
google-cloud-storage = "^2.13.0"
firebase-admin = "^6.2.0"
pydantic = "^2.4.2"
opencv-python = "^4.8.1.78"
pillow = "^10.1.0"
firebase-functions = ">=0.1.0,<0.2.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79